name change: preprocessing>> create_everything
I decided to make this file generate everything that needs to be loaded and or changed for the model and its comparison. So it will be loading the tokenizer, two instances of the pretrained model (1 for comparison and 1 for custom model), and the datasets. This will add the new language tokens to the tokenizer and then save it locally. Next it will refit both model instances' embedding sizes and save the two model checkpoints under different names. Finally it will download all the needed datasets, split them, preprocess them, combine the splits into a train and test dataset, and save them locally for ease of loading.